First draft of structure based threadqueues
Instead of compiling in thread queues, let's treat thread queues as
structures of function pointers, a bit like Linux's pluggable
schedulers. This lets us pick our schedulers at runtime (via an
environment variable?) and lets us get rid of tons of boilerplate (null
functions in the function pointer should be handled in a sane default
way in the code, this will take some work).

This is a partial draft of the work. I will keep committing it as I go.